+++ router/main.go
@@ -2,14 +2,8 @@ package router
 
 import (
 	"embed"
-	"fmt"
-	"net/http"
-	"os"
-	"strings"
 
 	"github.com/gin-gonic/gin"
-	"github.com/songquanpeng/one-api/common/config"
-	"github.com/songquanpeng/one-api/common/logger"
 	"github.com/songquanpeng/one-api/middleware"
 )
 
@@ -25,17 +19,5 @@ func SetRouter(router *gin.Engine, buildFS embed.FS) {
 	SetApiRouter(router)
 	SetDashboardRouter(router)
 	SetRelayRouter(router)
-	frontendBaseUrl := os.Getenv("FRONTEND_BASE_URL")
-	if config.IsMasterNode && frontendBaseUrl != "" {
-		frontendBaseUrl = ""
-		logger.SysLog("FRONTEND_BASE_URL is ignored on master node")
-	}
-	if frontendBaseUrl == "" {
-		SetWebRouter(router, buildFS)
-	} else {
-		frontendBaseUrl = strings.TrimSuffix(frontendBaseUrl, "/")
-		router.NoRoute(func(c *gin.Context) {
-			c.Redirect(http.StatusMovedPermanently, fmt.Sprintf("%s%s", frontendBaseUrl, c.Request.RequestURI))
-		})
-	}
+	SetWebRouter(router, buildFS)
 }